attach 변수 이름, size 타입바꿈

+++ cookies/src/main/java/com/mjc813/cookies/models/attach/Attach.java
@@ -10,8 +10,8 @@ public interface Attach {
     String getFileName();
     void setFileName(String fileName);
 
-    Long getSize();
-    void setSize(Long size);
+    Integer getSize();
+    void setSize(Integer size);
 
     String getExt();
     void setExt(String ext);
